+++ src/cli/run.rs
@@ -1,8 +1,11 @@
-use crate::prelude::{
-    AppError, Storage,
-    command::{Cli, Commands, SortKey},
-    contact::{Contact, ValidationReq, phone_number_matches},
-    file::load_migrated_contact,
+use crate::{
+    domain::contact,
+    prelude::{
+        AppError, Storage,
+        command::{Cli, Commands, SortKey},
+        contact::{Contact, ValidationReq, phone_number_matches},
+        file::load_migrated_contact,
+    },
 };
 
 use clap::Parser;
@@ -35,7 +38,7 @@ pub fn run_app() -> Result<(), AppError> {
                 phone,
                 email.unwrap_or_default(),
                 tag.unwrap_or_default(),
-            )?;
+            );
 
             if !new_contact.validate_name()? {
                 return Err(AppError::Validation(ValidationReq::name_req()));
@@ -123,6 +126,44 @@ pub fn run_app() -> Result<(), AppError> {
                 );
             }
 
+            Ok(())
+        }
+
+        // Edit Contact
+        Commands::Edit {
+            name,
+            phone,
+            new_name,
+            new_phone,
+            new_email,
+            new_tag,
+        } => {
+            let desired_contact = Contact::new(name, phone, "".to_string(), "".to_string());
+            let found_contact = storage
+                .mem_store
+                .data
+                .iter_mut()
+                .find(|c| **c == desired_contact);
+
+            if let Some(contact) = found_contact {
+                if let Some(name) = new_name {
+                    contact.name = name;
+                }
+                if let Some(phone) = new_phone {
+                    contact.phone = phone;
+                }
+                if let Some(email) = new_email {
+                    contact.email = email;
+                }
+                if let Some(tag) = new_tag {
+                    contact.tag = tag;
+                }
+
+                contact.updated_at = Some(contact::Utc::now());
+            }
+
+            storage.save()?;
+            println!("Contact updated successfully");
             Ok(())
         }